+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 - PIM-9593: Prevent job "compute project calculation" from appearing in the process tracker and in the dashboard
 - PIM-9594: Fix project completeness when at least 1 attribute group have the "All" group in its permissions but the project's locale permissions does not
 - PIM-9600: Fix rules that use string concatenation with measurement ending by zero
+- PIM-9606: Fix rule execution job status when stopping the job during the last rule
 
 ## Improvements